requirements.txt
langchain==0.2.17
langchain-core==0.2.43
langgraph==0.2.14
langchain-groq==0.1.9

chromadb==0.5.5

sentence-transformers==2.7.0
torch>=2.0.0

streamlit==1.37.1

tqdm
numpy
python-dotenv